src/ServerPulse.Frontend/src/app/modules/server-slots/services/server-slot-data-collector/server-slot-data-collector.service.ts
import { Injectable } from '@angular/core';
import { Observable } from 'rxjs';
import { SlotDataApiService, SlotDataResponse } from '../../../shared';
import { ServerSlotDataCollector } from './server-slot-data-collector';

@Injectable({
  providedIn: 'root'
})
export class ServerSlotDataCollectorService implements ServerSlotDataCollector {

  constructor(
    private readonly apiService: SlotDataApiService
  ) { }

  getServerSlotData(slotKey: string): Observable<SlotDataResponse> {
    return this.apiService.getSlotData(slotKey);
  }
}
